Move remote mount logic to a file of its own
